requirement.txt
streamlit
pandas
python-dotenv
google-generativeai
langchain
